We have done a lot since the 2 yars since we have updated this... we updated missiles, sans + bones, Handsome Yoshi Easter Egg, +Lives, +Shield, +B_O_B & More

+++ FirstEverGame!/src/Game_Object.java
@@ -3,30 +3,30 @@ import java.awt.Graphics;
 import java.awt.Rectangle;
 
 public class Game_Object {
-	int x;
-	int y;
+	double x;
+	double y;
 	int width;
 	int height;
-	int speed = 10;
+	double speed = 10;
 	boolean isAlive = true;
 	Rectangle collisionBox;
-	Game_Object(int x, int y, int width, int height) {
+	Game_Object(double x, double y, int width, int height) {
 		this.x = x;
 		this.y = y;
 		this.width = width;
 		this.height = height;
 		this.isAlive = true;
-		this.collisionBox = new Rectangle(x,y,width,height);
+		this.collisionBox = new Rectangle((int) x, (int) y, width, height);
 	}
 
 	public void draw(Graphics g) {
 		g.setColor(Color.GRAY);
-		g.fillRect(x, y, width, height);
+		g.fillRect((int) x, (int) y, width, height);
 	
 	}
 	
 	public void update(){
-		collisionBox.setBounds(x, y, width, height);
+		collisionBox.setBounds((int) x, (int) y, width, height);
 	}
 	
 }
